Context in context (#2745)
* context in context

* keeping comments

* minimal context window line

* nit
