+++ .changeset/curvy-walls-cry.md
@@ -1,0 +1,5 @@
+---
+"claude-dev": minor
+---
+
+Added context window usage to env variables context for messages
